+++ sqllin-driver/src/commonMain/kotlin/com/ctrip/sqllin/driver/DatabaseConnection.kt
@@ -35,11 +35,5 @@ public interface DatabaseConnection {
 
     public fun close()
 
-    @Deprecated(
-        message = "The property closed has been deprecated, please use the isClosed to replace it",
-        replaceWith = ReplaceWith("isClosed")
-    )
-    public val closed: Boolean
-
     public val isClosed: Boolean
 }